Move view helpers to views file
